Remove NVSHMEM, dead code, and fix broken tests
- Remove _NvshmemBufferPool and NVSHMEM detection (private API
  torch.distributed._symmetric_memory). _alltoall_v now uses NCCL only.
- Remove unused functions: _alltoall_v_nccl, _alltoall_v_async,
  _return_lse, _scatter_dkv_to_owners, preinit_nvshmem_buffers
- Remove unused ring-pass functions: _compute_step_ffa,
  _compute_and_merge_step, _backward_step_ffa, _backward_step
- Fix test imports: remove references to deleted functions, move
  ring-pass test helpers into test file
- Remove TestExtractTokensForDocs and TestScatterToChunk (tested
  functions that no longer exist)
