+++ torchtitan/distributed/varlen_cp/ring_attention.py
@@ -11,7 +11,7 @@ Based on: Men et al., "Magi Attention: Efficient Attention Mechanism for
 Extreme-Scale LLM Training", https://arxiv.org/abs/2505.13211
 
 Entry point: varlen_ring_attention() delegates to varlen_magi_dispatch().
-Also provides shared utilities: FFA kernel wrappers, LSE merge, backward steps.
+Also provides shared utilities: LSE merge, cu_seqlens helpers.
 """
 
 from __future__ import annotations
@@ -19,14 +19,6 @@ from __future__ import annotations
 import torch
 from torch.distributed.device_mesh import DeviceMesh
 from torchtitan.distributed.varlen_cp.dispatch_solver import DispatchPlan
-
-# PyTorch-native flex_attention wrappers for range-based attention.
-# These replace the external magi-attention FFA kernel dependency with
-# torch.nn.attention.flex_attention (compiled Triton kernel).
-from torchtitan.distributed.varlen_cp.flex_attn_kernels import (
-    flex_attn_backward,
-    flex_attn_forward,
-)
 
 
 # ---------------------------------------------------------------------------
@@ -171,284 +163,6 @@ def _build_ring_step_cu_seqlens(
     cu_seqlens_k = torch.tensor(cu_k_list, dtype=torch.int32, device=device)
 
     return cu_seqlens_q, cu_seqlens_k, q_doc_ranges, k_doc_ranges, num_real_q_tokens, is_causal
-
-
-
-# ---------------------------------------------------------------------------
-# FFA (Flex-Flash-Attention) compute path
-# ---------------------------------------------------------------------------
-
-
-def _compute_step_ffa(
-    q: torch.Tensor,
-    cur_k: torch.Tensor,
-    cur_v: torch.Tensor,
-    q_doc_ranges: list[tuple[int, int]],
-    k_doc_ranges: list[tuple[int, int]],
-    is_causal: bool,
-    q_chunk_start: int,
-    k_chunk_start: int,
-    chunk_size: int,
-    n_heads: int,
-    head_dim: int,
-    accum_out: torch.Tensor,
-    accum_lse: torch.Tensor,
-) -> tuple[torch.Tensor, torch.Tensor]:
-    """Compute one (Q,K) step using magi-attention's FFA kernel and merge.
-
-    Operates on full chunk-level tensors with range descriptors.  Each doc's
-    Q/K intersection becomes one range pair for flex_attention.
-
-    Args:
-        q: Full Q chunk, shape (chunk_size, n_heads, head_dim).
-        cur_k: Full K chunk, shape (chunk_size, n_kv_heads, head_dim).
-        cur_v: Full V chunk, shape (chunk_size, n_kv_heads, head_dim).
-        q_doc_ranges: Global (start, end) pairs for Q docs.
-        k_doc_ranges: Global (start, end) pairs for K docs.
-        is_causal: True for diagonal blocks, False for below-diagonal.
-        q_chunk_start: Global start of the Q chunk.
-        k_chunk_start: Global start of the K chunk.
-        chunk_size: Tokens per chunk.
-        n_heads: Number of Q attention heads.
-        head_dim: Dimension per head.
-        accum_out: Running accumulated output.
-        accum_lse: Running accumulated LSE.
-
-    Returns:
-        Updated (accum_out, accum_lse).
-    """
-    device = q.device
-
-    # Convert global doc ranges to chunk-local coordinates for FFA
-    q_ranges = torch.tensor(
-        [[qs - q_chunk_start, qe - q_chunk_start] for qs, qe in q_doc_ranges],
-        dtype=torch.int32,
-        device=device,
-    )
-    k_ranges = torch.tensor(
-        [[ks - k_chunk_start, ke - k_chunk_start] for ks, ke in k_doc_ranges],
-        dtype=torch.int32,
-        device=device,
-    )
-    # MaskType mapping: 0=FULL, 1=CAUSAL — matches FFA's attn_type_map
-    attn_type = 1 if is_causal else 0
-    attn_type_map = torch.full(
-        (len(q_doc_ranges),), attn_type, dtype=torch.int32, device=device
-    )
-
-    step_out, step_meta = flex_attn_forward(
-        q,
-        cur_k,
-        cur_v,
-        q_ranges,
-        k_ranges,
-        attn_type_map,
-    )
-
-    # flex_attn LSE shape: (chunk_size, n_heads) float32 -> (n_heads, chunk_size)
-    step_lse = step_meta.lse.transpose(0, 1)
-
-    # Positions outside q_ranges have undefined output/LSE from FFA.
-    # Mask them to -inf LSE / zero output so they don't affect the merge.
-    q_coverage = torch.zeros(chunk_size, dtype=torch.bool, device=device)
-    for qs, qe in q_doc_ranges:
-        q_coverage[qs - q_chunk_start : qe - q_chunk_start] = True
-    uncovered = ~q_coverage
-    if uncovered.any():
-        step_lse[:, uncovered] = float("-inf")
-        step_out[uncovered] = 0
-
-    return merge_with_lse(accum_out, accum_lse, step_out, step_lse)
-
-
-# ---------------------------------------------------------------------------
-# Shared per-step attention helper
-# ---------------------------------------------------------------------------
-
-
-def _compute_and_merge_step(
-    q: torch.Tensor,
-    cur_k: torch.Tensor,
-    cur_v: torch.Tensor,
-    global_cu_seqlens: torch.Tensor,
-    q_chunk_start: int,
-    q_chunk_end: int,
-    k_chunk_start: int,
-    k_chunk_end: int,
-    original_total_seqlen: int,
-    chunk_size: int,
-    accum_out: torch.Tensor,
-    accum_lse: torch.Tensor,
-) -> tuple[torch.Tensor, torch.Tensor]:
-    """Compute partial attention for one (Q_chunk, K_chunk) pair and merge.
-
-    Builds cu_seqlens for the (Q,K) pair and calls flex_attention via
-    ``_compute_step_ffa``, merging into the running accumulator.
-
-    Args:
-        q: Local Q chunk, shape (chunk_size, n_heads, head_dim).
-        cur_k: K for this step, shape (chunk_size, n_kv_heads, head_dim).
-        cur_v: V for this step, shape (chunk_size, n_kv_heads, head_dim).
-        global_cu_seqlens: Global cumulative sequence lengths.
-        q_chunk_start: Global start of Q chunk.
-        q_chunk_end: Global end of Q chunk.
-        k_chunk_start: Global start of K chunk.
-        k_chunk_end: Global end of K chunk.
-        original_total_seqlen: Total sequence length before padding.
-        chunk_size: Chunk size.
-        accum_out: Running accumulated output.
-        accum_lse: Running accumulated LSE.
-
-    Returns:
-        Updated (accum_out, accum_lse).
-    """
-    device = q.device
-    dtype = q.dtype
-    n_heads = q.shape[1]
-    head_dim = q.shape[2]
-
-    (
-        cu_q,
-        cu_k,
-        q_doc_ranges,
-        k_doc_ranges,
-        num_real_q,
-        is_causal,
-    ) = _build_ring_step_cu_seqlens(
-        global_cu_seqlens,
-        q_chunk_start,
-        q_chunk_end,
-        k_chunk_start,
-        k_chunk_end,
-        original_total_seqlen,
-        chunk_size,
-        device,
-    )
-
-    if num_real_q == 0 or not k_doc_ranges:
-        return accum_out, accum_lse
-
-    return _compute_step_ffa(
-        q,
-        cur_k,
-        cur_v,
-        q_doc_ranges,
-        k_doc_ranges,
-        is_causal,
-        q_chunk_start,
-        k_chunk_start,
-        chunk_size,
-        n_heads,
-        head_dim,
-        accum_out,
-        accum_lse,
-    )
-
-
-# ---------------------------------------------------------------------------
-# Backward helper (flex_attention backward kernel)
-# ---------------------------------------------------------------------------
-
-
-def _backward_step_ffa(
-    grad_output: torch.Tensor,
-    q: torch.Tensor,
-    cur_k: torch.Tensor,
-    cur_v: torch.Tensor,
-    merged_out: torch.Tensor,
-    merged_lse: torch.Tensor,
-    global_cu_seqlens: torch.Tensor,
-    q_chunk_start: int,
-    q_chunk_end: int,
-    k_chunk_start: int,
-    k_chunk_end: int,
-    original_total_seqlen: int,
-    chunk_size: int,
-    grad_q: torch.Tensor,
-    grad_kv: torch.Tensor,
-) -> None:
-    """Compute backward for one (Q,K) step using flex_attention backward.
-
-    Operates on full chunk-level tensors with range descriptors (no pack/unpack).
-
-    Args:
-        Same as ``_backward_step``.
-    """
-    device = q.device
-    n_kv_head_dim = cur_k.shape[2]
-
-    cu_q, cu_k, q_doc_ranges, k_doc_ranges, num_real_q, is_causal = (
-        _build_ring_step_cu_seqlens(
-            global_cu_seqlens,
-            q_chunk_start, q_chunk_end,
-            k_chunk_start, k_chunk_end,
-            original_total_seqlen, chunk_size, device,
-        )
-    )
-
-    if num_real_q == 0 or not k_doc_ranges:
-        return
-
-    # Convert doc ranges to FFA format (chunk-local coordinates)
-    q_ranges = torch.tensor(
-        [[qs - q_chunk_start, qe - q_chunk_start] for qs, qe in q_doc_ranges],
-        dtype=torch.int32, device=device,
-    )
-    k_ranges = torch.tensor(
-        [[ks - k_chunk_start, ke - k_chunk_start] for ks, ke in k_doc_ranges],
-        dtype=torch.int32, device=device,
-    )
-    attn_type = 1 if is_causal else 0
-    attn_type_map = torch.full(
-        (len(q_doc_ranges),), attn_type, dtype=torch.int32, device=device,
-    )
-
-    # LSE shape: (n_heads, chunk_size) -> (chunk_size, n_heads) for backward
-    ffa_lse = merged_lse.transpose(0, 1).contiguous()
-
-    dq, dk, dv = flex_attn_backward(
-        grad_output.contiguous(),
-        q.contiguous(),
-        cur_k.contiguous(),
-        cur_v.contiguous(),
-        merged_out.contiguous(),
-        ffa_lse,
-        q_ranges,
-        k_ranges,
-        attn_type_map,
-    )
-
-    # Accumulate into chunk-level gradient buffers
-    grad_q += dq
-    grad_kv[..., :n_kv_head_dim] += dk
-    grad_kv[..., n_kv_head_dim:] += dv
-
-
-def _backward_step(
-    grad_output: torch.Tensor,
-    q: torch.Tensor,
-    cur_k: torch.Tensor,
-    cur_v: torch.Tensor,
-    merged_out: torch.Tensor,
-    merged_lse: torch.Tensor,
-    global_cu_seqlens: torch.Tensor,
-    q_chunk_start: int,
-    q_chunk_end: int,
-    k_chunk_start: int,
-    k_chunk_end: int,
-    original_total_seqlen: int,
-    chunk_size: int,
-    grad_q: torch.Tensor,
-    grad_kv: torch.Tensor,
-) -> None:
-    """Compute backward for one (Q,K) step using flex_attention backward."""
-    _backward_step_ffa(
-        grad_output, q, cur_k, cur_v, merged_out, merged_lse,
-        global_cu_seqlens,
-        q_chunk_start, q_chunk_end, k_chunk_start, k_chunk_end,
-        original_total_seqlen, chunk_size,
-        grad_q, grad_kv,
-    )
